+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- added : icon and title for page internal view
 - changed : rename history to evolution
 - changed : make jsonjsdb_editor use xlsx state for evolution
 - fixed : favorite page about tab image caption